Added necessary html code to generate the new style SDI pages.
The table attributes variable has been changed to remove the html
border. The javascript paint function do the job better.

The starttable function call has been changed to accept a Title
argument, that will be atached to the table top bar.

The html code inside starttable() is not the best way to write,
but it's easier to maintain, as we can't wrap lines due to html
specific issues.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>