Fixed problems with menu expand action on pages with multiples tables.
Before, we use just one global variable to prevent the menu to be showed
when the table is expanded down. With multiples tables, this was a problem
cause if one table is expanded the others tables columns menu should open
normaly.

Now, this has been fixed using a bidimensional array to represent each
table current state.

Signe-ooff-th menu expand action on pages with multiples tables.