Added the SDI bar support to the sumary page.
Before, the sumary was a single page with some important information.

Now, the bar give a new way to navegate thru the SDI generated pages.
Note that we added a div element with the pagetype on it. This is essential
to javascript know where to send the URL when selected by the select box.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>