Fixed the way generatebar give the result.
Before, we was using a file to hold the generated bar during the
generating progress. This isn't good because if more than one stances
of generator.sh are running, the bar that pages will grab is undefined.

Now, we created a function that return the HTML content of the SDI bar.
So, then program that use the bar have a way to hold the value during all
generating proccess.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>