sdiweb: Created generateclasspage.sh to generate classes pages.
When launching sdi, we'll create all web pages. This script provides the
creation of a class page for a single host. A class page is the main page
of a class, keeping all hosts in a organized table.

At this point, we use SSI server tecnology to easy mantain our web pages.
The page header, footer and tables comes from models files.

This script assumes that SDIBAR variable is defined and have the bar html.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Acked-by: Bruno Ribas <ribas@c3sl.ufpr.br>