sdiweb: Created generatesdibar.sh to generate sdibar code.
Based on CLASSESDIR folder, the script gets the pages avaliable to sdi
users and returns the html needed to create the sdi bar.

For this generation process, a model file for sdi bar is used to keep
an easy way to modify the bar.

The main function of script is create the selectbox that allows navegate
truh the sdi classes page.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Acked-by: Bruno Ribas <ribas@c3sl.ufpr.br>