sdiweb: Created generatexmls.sh to generate XML files.
The XML files with SSI tecnology will be the main reason because sdi
becomes a real instant diagnose system.

This script is the creator of XML-SSI integration. Our XML files have
informations about the hosts, a tag for each script running.

What happens it's that when users request the XML file, the server will
include others files to complete the file, and those files will be always
updated (provided by sdi tunnel).

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Acked-by: Bruno Ribas <ribas@c3sl.ufpr.br>