launchsdi.sh becomes the SDI launcher
Since SDI 0.0.99, which was our Proof of Concept implementation, no formal
sdi tunnel launcher existed and were left to the administrator to create a
script to do so. PRDSDI is the example of that, and it were not generic.

As previously rename of launchsdi.sh to launchsditunnel.sh, this commit
modifies launchsdi.sh to be like old PRDSDI but more generic.  It includes
functions to generate class structure, to create WWWDIR folders and links,
to retrieve enabled columns to web pages and also includes a complete host
search through CLASSDIR folder.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>