Fix 020-raid to identify partitions out of the raid and synchronizing
Disks that are synchronizing will be identified as
${DISK}${PARTITION}_SYNCHRONIZING_RAID and partitions that are out of the
raid will be identified as $DISK${PARTITION}_OUT_OF_RAID

These new features to 020-raid will make the job of identifying problems
easier as we will be able to build a more specific summary. Also it will be
able to identify raid partitions that keep synchronizing all the time, which
may mean that the hard disk table is corrupted.

Signed-off-by: Luis Lourenço <lhal04@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>