sdiweb: Added generation of summaries XML files
The summary have a special XML format. The function generatesummariesxml()
will generate this XML to be loaded by summary page.

The summary configuration file must have two elements: SNAME is the
summary name, used on generatesdibar script. And a vector STATES that have
all states that should be showned by the summary.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>