Rewrite sources idea in PARSE() to decrease the disk usage during execution
When running a PARSE function to each HOST on SDI, the number of sources on
script files are one per message. If there's 2000 hosts with 10 diagnose
scripts running hour by hour, the number of sources will be 20000 per hour.

With this commit we only source the parser object once per host, and on
demand.  For example, with 2000 hosts and 10 scripts, the first time a line
comes to SDI, it will source the parser object but in the next hour this
source will not be necessary anymore. This will reduce 20000 disk access on
each hour.

Also, a trap has been added to force a re-source of parser objects. This
could be useful if some parser object has been updated and you don't want to
reopen all SDI tunnels again. On the next time the script line comes to SDI
PARSE() the source will be made. This action can be done running
lauchsditunnel with option --reload-po.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>