Fix premature calling of closesdiprocs()
When killing all hosts on SDI the "exit 0" command is sent twice, the
first will kill the ssh tunnel and the other will cause a broken pipe,
killing the tail's.

The closesdiprocs() kills all pocesses related to SDI management, such
as the states fifo read. This cause no more mesages to be interpreted
by savestate.

Calling closesdiprocs() before wait the proccess cause the PARSE() to be
stalled when sending mesagens to SFIFO, witch reading is alreadly stoped.

The fix move closesdiprocs() after waitend().

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>