Added interval between exit commands
When killing all hosts on SDI the "exit 0" command is sent twice, the
first will kill the ssh tunnel and the other will cause a broken pipe,
killing the tail's.

But with 2 consecutives exit commands, theres no time to bash detect the
broken pipe, stalling the proccess.

This fix will provide a time between those commands. The minimum wait time
is one second (the tail update period), but 15 seconds are a good choice
to secure slow connections to safely closure.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>