Fix background tail alive after killing all hosts
When running the ssh connection receiving two tail's of commands, the
first one was on background. The bug happens when the "exit 0" command
is sent to CMDGENERAL file, because this tail will suffer a broken pipe,
but the first not.

This fix replace those tail's by just one, adding the '-q' option.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>