Fix savestate tail not be killed before it ends
When killing SDI processes an "exit" command is sent to SFIFO (which is the
FIFO of the running savestate), but after sending the 'exit' command all SDI
preccesses were killed (SIGKILLed) with no mercy and not properly finished.
Before killing this tail we need to wait savestate() to finish correctly, so
the remaining unprocessed states will be updated.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>