Fix failure when blocking not enabled scripts
When running SDI some scripts are available, and they can be enabled or not.
When sending a script to a HOST with a not enabled script, an error (that no
such parser object exist or is not enabled) is triggered and the HOST log
file will have this information.

The problem occurs when the second time the not enabled script is sent to
some HOST. The PARSE() detects at the first time, but when the message comes
again, the script functions were already sourced and the PARSE() will
execute the code.

When detecting a not enabled script functions previewsly loaded are unseted
so next time same script is executed at HOSTS, PARSE() will not log them as
an enabled-script.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>