launchsditunnel.sh: Communicate with the PARSE() through a fifo
Send messages from the ssh to the PARSE() using a fifo create by mkfifo(1),
instead of a shell pipe.
This change decreases drastically the number of process running in the
system, once a unique instance of PARSE() can read all messages comming
from the fifo.

The closesdiprocs() uses closefifo() function to close the PARSE() fifo.

Signed-off-by: Diego Giovane Pasqualin <dgp06@c3sl.ufpr.br>