Allow the user to set how many PARSE() instances must be created
The previous commit change the communication between ssh connection and
PARSE() function and implies on a creation of a unique PARSE() instance.
This is better for decrease the load of system, but may be worse to
performance.
This commit implements the possibility of the user manage this situation,
declaring the $HOSTSPERPARSE variable, who defines how many instances of
parse must be created.

Signed-off-by: Diego Giovane Pasqualin <dgp06@c3sl.ufpr.br>