Move web pages generation process to sdiweb
When launching SDI the class specific web files are created just before
launch all connections to this class. What happens is that some tunnels
could have already launched, decreasing the performance of web files
creation.

Moving all the creation proccess before all tunnels launching will decrease
the launching time. This is a concept change too.

Some tests with a empty machine on 74 hosts showed a little better
performance, but we expect that on a loaded machine with more hosts the
speed will be significantly improved.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Bruno Ribas <ribas@c3sl.ufpr.br>