Fix SDITUNNEL() to correctly work when ssh fails
We were using tail with --pid=$SDITUNNELPID, but when ssh fails the tail
still there, stacking the function until SDI receive a close command.

With this tail on background the script works normally, but it's necessary
to kill the tail after ssh exits.

This commit fix a regression added by commit 916fd6ef43c.