New method to control the number of PARSE() instances running
This method use a semaphore implemented with socket's to limit the number
of PARSE() instances running at the same time. This control intends to
decrease the load of the server.

The new variable "socket port", defined as 19183 in sdi.conf and
configsdiparser.py, configure the port to socketdaemon.py connect and
socketclient send commands.

The socketdaemon.py accept these commands:
    "acquire": request to acquire a slot in semaphore
    "release": release a slot in semaphore
    "status": show status of connections, including the number of clients
        sleeping in semaphore, the maximum number of simultaneous
        connections, and the current number of connections.
    "add": increase the maximum number of connections acceptable
        simultaneously
    "rem": decrease the maximum number of connections acceptable
        simultaneously
    "stop": close socket and exit program