Rewrite savestate() idea to reduce sources on states files
When running one savestate function the number of sources on states files
are one per state set. If there's 2000 hosts with one script that every hour
set six states, the number of sources will be 12000 per hour.

With this commit we only source the state file once on SDI, and on demand.
For example, with 2000 hosts and one script that set six states, the first
time a state set request comes to savestate, it will source the state file
but in the next hour this source will not be necessary anymore. This will
reduce 12000 disk access on each hour.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>