Fixed summaries columns selection
When selecting columns to be displayed on a summary table, all the table was
moved below the selection menu. This was happening because all the page
structure was being created via javascript, what was (by a unknown reason)
causing this problem.

Moving the table structure creation process to generatesummary script, the
problem was solved. That javascript code better now, because it will create
less document elements.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>