Added support to allow user to specify a fast directory to data
On configsdiparser.py some new variables has been added to allow the system
administrator to specify a fast directory (usually the RAM memory).

The launchsdi and launchsditunnel scripts have been modified in order to
follow those variables defined. The launchsdi script just make sure the fast
directory are created and usable, and the launchsditunnel script will add a
new entry on crontab if the use of fast directory is enabled.

Also, a sync script has been added to syncronize data between fast data
directory and the data directory. This script will create the data
directories based on the content of "data history format" variable. With
this feature, the log rotation are automaticaly made.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>