launchsdi.sh: Calling launchsditunnel.sh defining the CLASS variable
We are calling launchsditunnel.sh with the CLASS variable definied,
which will contain the class that the hosts belongs to.
Also, we are defining the format HOST:SEND:RECEIVE in the class
files, where SEND and RECEIVE will be the send and
receive application used by the HOST under send and
receive dir, or blank, meaning that host will use the default ones.

Signed-off-by: Ricardo Tavares de Oliveira <rto07@c3sl.ufpr.br>
Signed-off-by: Diego Giovane Pasqualin <dgp06@c3sl.ufpr.br>