Changed the way we discover if a host belong to a state
Before, every state change we was using a grep to find out if a host was
on some state. This is very expensive.

The new way to do this, is to hold a folder to each state with empty files
in there, where the file name represent the host name. With this
modification we expect to get a better performance on states changes.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Diego Giovane Pasqualin <dgp06@c3sl.ufpr.br>