Fixed states.sh problem when relaunching sdi
The recreation of states fifo was being performed every time the states.sh
script was called. If some state daemon was already running, the fifo will
be recreating and the daemon will be stalled.

Changing the fifo recreation proccess to right before the real launch of the
states daemon will prevent this to happen.

Signed-off-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>
Signed-off-by: Diego Giovane Pasqualin <dgp06@c3sl.ufpr.br>