socket: Added a simple socket server interface
This socket interface connects to localhost and waits for connections on
port 18193 (for now). The objective is to provide an interface which can
be used by SDICORE application, replacing the socketdaemon.py.

The constructor method creates a socket to listen the port and the
GetMessage() waits for a client connection and returns a pointer to
a received message.

This commit is part of a series of commits which will become the SDICORE.

Signed-off-by: Diego Giovane Pasqualin <dgp06@c3sl.ufpr.br>