sdicore: Initial implementation of SDICORE
SDICORE is coming to replace the current method of process managment used
by SDI. Now it is done using forks (with one new PARSE() process for each
host opened), a separate system for load balancing and a complicated
function to ensure that all process created by SDI will be closed at the end
of the execution.

The SDICORE is being created to improve this, implementing a system
to schedule the incoming messages between a number of threads that fits
correctly the needs of the moment and can be managed much more precisely.

Signed-off-by: Diego Giovane Pasqualin <dgp06@c3sl.ufpr.br>