consumer: Fix possible eternal blocking on consumer thread
When a consumer thread is closed to save resources, it could wait for a
message comming from socket that could never come. So, these timeouts added
to sem_wait() (now sem_timedwait()) allow the execution flow to return to
sdicore, which can decide if the thread must close or not.

Signed-off-by: Diego Giovane Pasqualin <dgp06@c3sl.ufpr.br>
Acked-by: Ricardo T Oliveira <rto07@c3sl.ufpr.br>
Acked-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>