misc.sh: Added functions to manage cron to run SDI scripts
The function configure cron is responsable to configure the cron to run SDI
scripts to send commands to connected hosts at specific moments. The current
supported periods are minutely, hourly, daily, weekly and montly. It also
configures the data sincronization script to allow the use of fast data
directories to store incoming messages (and sincronize with reliable storage for
backup).

The function removecronconfig() is responsable to remove any SDI cron
configuration created during the startup proccess.

Signed-off-by: Vinicius Kwiecien Ruoso <vinicius.ruoso@gmail.com>