launchscripts.sh: Added bash protection against spaced paths
By adding double quotes to the paths used by the scripts we are adding support
to run them using paths that have spaces into them. We are also disabling the
possibility to run the launchscript command in the "onconnect" folder. This is
necessary until the boostrap command is not prepared to be sent more than once
to each host.

Signed-off-by: Vinicius Kwiecien Ruoso <vinicius.ruoso@gmail.com>