Updated related scripts to use webservice configurations
Updating the bootstrap command to rely on already set variables with information
on how to send the commands results to the webservice.

We are also changing the startsdi.sh script to pass this information to the
tunnel. Another feature added in the script is to set the hosts status as
OFFLINE when starting the tunnel and lose of connection.

Signed-off-by: Vinicius Kwiecien Ruoso <vinicius.ruoso@gmail.com>