misc.sh - Print function bug fix
The print function used the '>>' operator to write data to files. However, in some cases the file still did not exist, generating errors. The new code checks if the destination file exists, if it exists the function uses the >> operator, if the destination file does not exist, the function uses the '>' operator.